Move getMaxTimeAllowed to base class with schema field
